+++ cli/src/main/java/app/notesr/cli/command/Command.java
@@ -15,7 +15,7 @@ import java.util.List;
 import java.util.concurrent.Callable;
 
 import static app.notesr.cli.crypto.FileCryptor.KEY_GENERATOR_ALGORITHM;
-import static app.notesr.cli.util.PathUtils.getNameWithoutExtension;
+import static app.notesr.cli.util.FileUtils.getNameWithoutExtension;
 import static app.notesr.cli.util.Wiper.wipeDir;
 import static app.notesr.cli.util.Wiper.wipeFile;
