+++ cli/src/test/java/app/notesr/cli/command/CompileCommandTest.java
@@ -15,7 +15,7 @@ import java.util.Random;
 import static app.notesr.cli.command.Command.FILE_RW_ERROR;
 import static app.notesr.cli.command.Command.SUCCESS;
 import static app.notesr.cli.util.FixtureUtils.getFixturePath;
-import static app.notesr.cli.util.PathUtils.getNameWithoutExtension;
+import static app.notesr.cli.util.FileUtils.getNameWithoutExtension;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
